requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
bcrypt==4.0.1
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comtypes==1.1.14
constantly==15.1.0
cron-descriptor==1.2.35
crypto==1.4.1
cryptography==39.0.2
cssselect==1.2.0
cssutils==2.6.0
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dj-static==0.0.6
Django==4.1.7
django-admin-rangefilter==0.9.0
django-celery-beat==2.4.0
django-import-export==3.1.0
django-pipeline==2.1.0
django-timezone-field==5.0
djangorestframework==3.14.0
dnspython==2.3.0
emails==0.6
ems==0.2.25
et-xmlfile==1.1.0
exceptiongroup==1.1.1
Flask==2.2.3
flatbuffers==23.3.3
frozenlist==1.3.3
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
h5py==3.8.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
hyperlink==21.0.0
ibm-cloud-sdk-core==3.16.2
ibm-watson==7.0.0
idna==3.4
importlib-metadata==3.10.1
incremental==22.10.0
itsdangerous==2.1.2
jax==0.4.6
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kombu==5.2.4
ldap3==2.9.1
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.2
marshmallow==3.19.0
mongoengine==0.27.0
msgpack==1.0.5
multidict==6.0.4
mysqlclient==2.1.1
Naked==0.1.32
netunnel==1.0.11
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
odfpy==1.4.1
openai==0.27.2
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.1
paramiko==3.1.0
premailer==3.10.0
prompt-toolkit==3.0.38
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pymongo==4.3.3
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pypiwin32==223
PySocks==1.7.1
pyTelegramBotAPI==4.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-telegram-bot==20.1
python3-openid==3.2.0
pyttsx3==2.90
pytz==2022.6
pywin32==305
PyYAML==6.0
redis==4.3.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scipy==1.10.1
selenium==4.8.3
service-identity==21.1.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.1.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.4.0
social-auth-storage-mongoengine==1.0.1
sortedcontainers==2.4.0
SpeechRecognition==3.10.0
sqlparse==0.4.3
sshtunnel==0.4.0
static3==0.7.0
tablib==3.3.0
telebot==0.0.5
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0rc1
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0rc1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.6
websocket-client==1.1.0
Werkzeug==2.2.3
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
zipp==3.15.0
zope.interface==6.0
